requirements.txt
pylint==2.16.4
flask==2.2.3
gunicorn==20.1.0
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
python-dotenv==1.0.0
flask-restful==0.3.9
flask-mysqldb==1.0.1
coverage==7.2.1
requests==2.28.2
pytest==7.2.2
